show parent sidebar if parent is yoast/job-posting or yoast/recipe
(allows testing to go forward)
